Move Board to components dir
